+++ packages/feedback-react/src/index.ts
@@ -1,3 +1,2 @@
 export { Feedback } from "./Feedback";
-export { SimplifiedFeedback } from "./SimplifiedFeedback";
-export { VERY_UNHAPPY, UNHAPPY, NEUTRAL, HAPPY, VERY_HAPPY } from "./FeedbackValues";
+export { PRESETS } from "./presets";